requirements.txt
absl-py==2.1.0
astunparse==1.6.3
cachetools==5.3.2
certifi==2024.2.2
charset-normalizer==3.3.2
colorama==0.4.6
contourpy==1.1.1
cycler==0.12.1
filelock==3.13.1
fire==0.5.0
flatbuffers==23.5.26
fonttools==4.48.1
fsspec==2024.2.0
gast==0.4.0
google-auth==2.27.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
grpcio==1.60.1
h5py==3.10.0
idna==3.6
imageio==2.34.0
importlib-metadata==7.0.1
importlib-resources==6.1.1
Jinja2==3.1.3
joblib==1.3.2
keras==2.13.1
kiwisolver==1.4.5
lazy_loader==0.3
libclang==16.0.6
Markdown==3.5.2
MarkupSafe==2.1.5
matplotlib==3.7.4
medmnist==3.0.1
mpmath==1.3.0
networkx==3.1
numpy==1.24.3
oauthlib==3.2.2
opt-einsum==3.3.0
packaging==23.2
paho-mqtt==1.6.1
pandas==2.0.3
pillow==10.2.0
protobuf==4.25.2
pyasn1==0.5.1
pyasn1-modules==0.3.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2024.1
PyWavelets==1.4.1
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.9
scikit-image==0.21.0
scikit-learn==1.3.2
scipy==1.10.1
six==1.16.0
sympy==1.12
tensorboard==2.13.0
tensorboard-data-server==0.7.2
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-intel==2.13.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.4.0
threadpoolctl==3.2.0
tifffile==2023.7.10
torch==2.2.0
torchvision==0.17.0
tqdm==4.66.1
typing_extensions==4.9.0
tzdata==2023.4
urllib3==2.2.0
Werkzeug==3.0.1
wrapt==1.16.0
xgboost==2.0.3
zipp==3.17.0
zlib-ng==0.4.0
